requirements.txt
bufferer==0.13
numpy==1.11.0
pandas==0.19.2
ffmpeg-normalize>=1.0.9
pyyaml==5.1
youtube_dl==2020.9.6
tabulate==0.8.7
paramiko==2.7.2
git+https://github.com/bitmovin/bitmovin-python.git@v1.21.0#egg=bitmovin-python
idna==2.5
urllib3==1.21.1
